+++ Sources/KeyboardKit/Gestures/DragGestureHandler.swift
@@ -15,56 +15,8 @@ import UIKit
  */
 public protocol DragGestureHandler {
     
-    func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint)
-}
-
-
-/**
- This drag gesture handler handles the space key cursor move
- drag gesture.
- */
-public class SpaceCursorDragGestureHandler: DragGestureHandler {
-    
-    public init(
-        context: KeyboardContext,
-        feedbackHandler: KeyboardFeedbackHandler,
-        sensitivity: SpaceDragSensitivity = .medium) {
-        self.context = context
-        self.feedbackHandler = feedbackHandler
-        self.sensitivity = sensitivity
-    }
-    
-    private let context: KeyboardContext
-    private let feedbackHandler: KeyboardFeedbackHandler
-    
-    public let sensitivity: SpaceDragSensitivity
-    
-    private var currentDragStartLocation: CGPoint?
-    private var currentDragTextPositionOffset: Int = 0
-    private var textDocumentProxy: UITextDocumentProxy { context.textDocumentProxy }
-    
     /**
-     Handle a drag gesture on space, which by default should
-     move the cursor left and right after a long press.
+     Handle drag gestures from a start to a current location.
      */
-    public func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
-        tryStartNewDragGesture(from: startLocation, to: currentLocation)
-        let dragDelta = startLocation.x - currentLocation.x
-        let textPositionOffset = Int(dragDelta / CGFloat(sensitivity.points))
-        guard textPositionOffset != currentDragTextPositionOffset else { return }
-        let offsetDelta = textPositionOffset - currentDragTextPositionOffset
-        textDocumentProxy.adjustTextPosition(byCharacterOffset: -offsetDelta)
-        currentDragTextPositionOffset = textPositionOffset
-    }
-}
-
-private extension SpaceCursorDragGestureHandler {
-    
-    func tryStartNewDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint) {
-        let isNewDrag = currentDragStartLocation != startLocation
-        currentDragStartLocation = startLocation
-        guard isNewDrag else { return }
-        currentDragTextPositionOffset = 0
-        feedbackHandler.triggerFeedbackForLongPressOnSpaceDragGesture()
-    }
+    func handleDragGesture(from startLocation: CGPoint, to currentLocation: CGPoint)
 }